Don't allow reschedule of ignition event to move stop time more than 90 degrees
This was originally 90 degrees for ignition, and the comment even says
90 degrees, but a few years ago this inexplicably was changed to 360
degrees. Aside from the fact that this seems wrong, the issue this was
originally intended to fix is cropping up:

Ignition pulses starting at the right time, but not releasing until a
full cycle later.  This results in misfires, and overdwelling, though
never a too-advance spark at least.

I'm not 100% sure this same problem couldn't happen to fueling pulses,
so for now, also adding the logic to fueling until more investigation
and tests are written.