Fix concurrent slot dma transfer safety check
The DMA is in mem-to-peripheral mode, which even in direct mode loads
the current transfer on the previous trigger, storing it in a single
transfer fifo.  This means that for the purpose of this safety check,
the transfer window that matters for the DMA is one time cycle before
its actual time value.

I don't love introducing a platform-specific weirdness here, but this is
the quick fix. We can look at making this more platform agnostic in the
future.