Simplify scheduler and eliminate direct DMA buffer manipulation
The old scheduler uses DMA double buffering to drive the GPIO, and does so with
a pair of 128 uS buffers representing the current 128 uS and the next 128 uS.
Substantial logic is used to allow these buffers to be changed while they are
the actively-used buffer. For example, since it is impossible to pause the dma,
we collect the time before and after the store into the buffer and explicitly
set the gpio if we believed we may have failed to set the bit, since it is
actually impossible to tell if we succeeded. This logic is error prone and very
difficult to validate. Many bugs have been found in the past around this logic:
even in direct mode, a single value is buffered by the DMA, and more recently,
being unable to rely on the stm32's DMA updating NDTR before triggering an
interrupt. Furthermore, all of this logic is fairly specific to the
limitations/abilities of the stm32f4 (#60)

One benefit of the current setup is that we can change upcoming events
immediately, even if they are <1 uS from occuring. If the buffer sizes are
reduced, this is an acceptable loss.

This change reduces the buffer size such that each buffer represents 32 uS, and
disallows changes to any buffers the DMA is using.

- `fired` and `scheduled` are replaced with a single state enum
- Entries that are unscheduled are always safe to schedule, though it can only
  schedule events after a certain point (between 32 uS - 64 uS from now)
- Entries that are scheduled can be "submitted" by the platform-dependent code.
  Any changes to scheduled events therefore must be with interrupts disabled.
- Entries that are submitted can not be changed, but it is unknown if they are
  fired yet.
- Entries that are fired are guaranteed to have fired, and can be safely marked
  unscheduled.

On an interval the scheduler changed scheduled events that are in the time range
of a buffer to submitted, and events that we fired as fired Once an event is
fired, and when both a start/stop are fired.  The platform code entirely owns
the process of submitting/marking-fired events, and hardware specifics such as
the dma ciccular buffers are entirely contained in the platform