Improve callback triggering mechanism.
First, add a helper for time_before_or_equal, and use it to solve a timing bug
where we miss an event that is at the current time.

Second, when scheduling a callback and determining it is already past due, we
immediately call the callback. This can result in unbounded time in interrupt,
and allows bugs to cause infinite recursion.  Instead we now asynchronously
pend the callback so that it is always triggered in the event timer interrupt
context.