Improve callback triggering mechanism.
First, add a helper for time_before_or_equal, and use it to solve a timing bug
where we miss an event that is at the current time.

Second, when scheduling a callback and determining it is already past due, we
immediately call the callback. This can result in unbounded time in interrupt,
and allows bugs to cause infinite recursion.  Instead we now asynchronously
pend the callback so that it is always triggered in the event timer interrupt
context.  Additionally, do not iterate over pending events in the
handler.  For a badly behaving callback function, this could
result in an infinite loop in the interrupt.  We no longer
need to rapidly handle back to back timer events with precision,
so just leave the interrupt pending, which will cause it to be
immediately handled.  This allows decoding to keep working.

Third, instead of exposing several primitives, expose one
new `schedule_event_timer` function that handles the single
use case of the timer callbacks.