table: separate 1d from 2d table types
This is implemented to save on ram, so we avoid needing a full
2d float table for things we know are 1d, and can never be 2d. This
change attempts to retain backwards compatibility: While the num-axis
field is removed, clients should be using the types in the structure
to know if a table is 1d or 2d.  Alternatively, the presence
of a vertical-axis field can be used instead.

1d tables continue to use the horizontal axis as their axis.