platform: factor out ADC handling and implement tlv2553
Like the max11632, the AD7888 and TLV2553 are split out into individual
header files that can be included. The TLC2543 is dropped.

This changes the stm32f4 sampling logic to match the gd32f4 (and stm32h7
 branch) where double buffered DMA is used for continuous sampling,
instead of triggering one batch of sampling at a time. On stm32f4, CS is
kept low constantly, and on gd32f4, CS is brought high between each sample.