sensors: refactor sensor api
For platform code that wants to report new sensor readings, there
is currently a single entrypoint `sensors_process` that takes the
sensor "source" type, and processes all of those. Instead, lets
provide independent entry points for each class of sensor. Here we
add `sensor_update_adc` and `sensor_update_freq`, and each take a
specific data structure representing, respectively, an ADC update,
or an individual frequency input's update.

This provides better separation between platform and sensor code,
as now the platform code is not directly writing into the `sensor_input`
structs.  If we use message queues one day, these structs directly
translate into the messages we would pass to a sensor handling
task.

Additionally, the sensor code should not aware of the number of bits
the ADC provides.  Remove the hardcoded 4096 values and require a
per sensor min/max raw value.  This also is necessary to properly
support frequency inputs, since we want linear mappings for those,
but a 12 bit value doesn't make any sense for them.

The platform must now provide raw sensor values in a float logical
form.  For the ADC values this is volts, for frequency inputs it is
hertz.  All values in configuration that reflect raw values must now
be floats and in the appropriate unit, rather than an integer directly
from the ADC.