scheduler: use tooth rpm for scheduling
This reverts a change from years ago to use the average rpm.  I think
the reasoning for the original change was to help with high rpm trigger
noise, but I can't see that the change makes much sense.  On the other,
as integration tests are showing, low rpm timing accuracy is greatly
compromised when not using the tooth rpm, such that accuracy tests will
not even pass.