integration: refactor integration framework to add support for hil tests
First, some refactoring of the sim framework to use types more thoroughly
and a cleaner interface: we translate all message, whether from the target
cbor console or test bench into a unified event type.

Then we split out the connector interface to support a simulation and
hardware-in-the-loop (HiL) connector.  This connector interfaces with the
viaems-fpga-tb software and hardware to control a real target viaems board
and collect the outputs, allowing the same python tests to run against real
hardware.

TODO add some documentation and references