console: implement cobs encoding for streaming console
This implements framing for the serial/USB communications, and additionally
adds a CRC32 to each message.  This allows a host communicating with the target
to scan for null delimiters instead of constantly trying to determine
if a message is valid.  Each CBOR payload has an HDLC CRC32 (Ethernet) appended,
and then the whole frame is COBS encoded and sent over the channel with NULL
bytes.

This is a breaking change for all clients.

This change also introduces the ability for the platform to bypass this API and directly
pass messages, which is useful for when the platform can provide native messages with
framing and error checking. This is intended to be used by Ethernet and CAN