Implement fixed-rate engine processing loop
This is a major re-architecture and refactor, with two main goals:
 - Move calculations and scheduling to not be triggered by the decoder,
   but at a fixed rate, and at a top-level location (viaems.c)
 - Remove as much state from individual modules as possible, and
   separate it from the global config structure

The former is done by separating out a function that does everything
needed for engine processing except for sensor and decoder calculations
to a dedicated function, `viaems_reschedule`, and having it be called by
the DMA buffer swap interrupt on stm32f4 and gd32f4.  The DMA double
buffers are increased to represent 200 uS (5 kHz), which is a compromise
to response latency, but simplifies things: All engine processing
happens at a fixed rate and is able to directly feed into populating the
upcoming schedule for the new buffer.  This is directly represented by
`struct platform_plan`, which is passed to `viaems_reschedule` and
expected to be populated with the list of upcoming events, as well as
changes to PWM and GPIO outputs during this time segment.  This has the
side benefit of removing significant amounts of the platform API, and
having all platform IO changes occur through a function call API,
allowing easier unit testing and benchmarking.

The latter state removal is done by splitting out decoder and sensor
configurations and states to their own structures, and having all module
functions take as parameters what they need.  Instead of relying on
global state structures for either, they are bundled together (along
with the state for output event scheduling, tasks, and calculations)
into a new `struct viaems`.  Since the only entrypoint to any of these
module functions (For engine calculations, etc) is now
`viaems_reschedule`, it can take a pointer to this structure and pass
state to modules as necessary.

The end result is that the platform code can now own the viaems state
structure. It must `viaems_init`, and then at an interval for
scheduling, call `viaems_reschedule`.  The console processing is moved
to `viaems_idle`, which the platform can call when it is done with
configuration of peripherals.  The only other place global access to
decoder/sensor state is needed is from interrupt handlers, and since the
platform already owns the state, it has access to it without needing any
API for it.

Long term, both decoding and sensor calculations should be moved into
`viaems_reschedule`, but since decoding can happen asynchronously -- in
fact, at high rpms, many decoder updates could happen in 200 uS -- it
would require some queueing of updates from the interrupt handlers.
This could be implemented with the newly-added spsc queue, or maybe via
an RTOS's queueing mechanism.

For now, both decoding and sensors are still handled in their respective
interrupt handlers, with decoding being higher priority, and sensor
processing equal to the DMA swaps, since they happen at the same rate.
All viaems code is now expected to not be called from multiple threads,
so it is on the platform to ensure it disables any interrupts as
necessary to prevent that from happening.  The benefit is that sensors,
decoding, calculations, etc do not need to have explicit critical
sections anymore.

The one exception is console processing, and this is solved by
implementing a queue between `viaems_reschedule` and `viaems_idle`'s
calling of `console_process` using a spsc queue.  Console feed messages
are now 1:1 with reschedule calls, so they are at a fixed 5 kHz rate.

Lastly, the benchmark is updated to take advantage of all the isolated
state. It produces friendlier output, and is generally far more useful
for evaluating a change's impact on performance reliably.