hosted: implement usb proxy for hil integration tests
The python usb connector implementation is not capable of
leveraging async usb, and thus cannot achieve the performance
needed to not drop event messages reliably.

Instead, we make a small proxy tool in C that connects
to the usb device and uses stdio just like the hosted-mode
build, which simplifies the python connector.  Later
we could also consider moving the rust tb interface
implementation into here as well, as to not depend on
a tool external to this repo.

fix