scheduler: improve handling of rescheduling events earlier
Currently, if a scheduled (though, not in progress) event is scheduled earlier
such that the start time has already passed, we give up on rescheduling it.
The end time of the event, or duration of the pulse if it is fuel, is preserved,
but this can result in heavily reduced dwell times.

Instead, when this happens, schedule the event to start as soon as possible and
correct the end time accordingly (if fuel).

To simplify the implementation, we remove the shared "safely" schedule function,
as it is a holdover from when we needed a critical section. Now handle fuel
and ignition totally separately, so its much more clear how each has special
handline for end angle vs duration.