calculations: add safety handling for overdwell and fuel overduty
Previously, the fuel scheduling allows for total disregard of fuel end angle
if it is needed to schedule, but this isn't desirable. It should only happen when
duty is high enough fudge values that much.  Instead, start failing
to schedule fuel events if we drift the end angle by more than 20 degrees.

This by itself can result in fuel events flapping in and out at the edge
of schedulability.  Add a max duty cycle configurable, and don't even
attempt fueling if we're above it, default is 95%.

For dwell, we do the same thing using our minimium fire(coil) time, and
introduce a configuration option so that the math can be aware
of wasted spark configurations.  However, we don't want to completely
disable ignition if we can't schedule our desired dwells: decrease
the dwell to fit up until we reach the configured minimum dwell.